Screen-Sync-Per-LED/requirements.txt
numpy
mss
pyserial
